+++ changelog.md
@@ -46,6 +46,7 @@ Codion Change Log
 - EntityExportModel.ForeignKeyNode.isCyclicalStub() renamed expandable().
 - EntityExportModel.Exporter added, DefaultEntityExportModel, minor refactoring.
 - EntityExportModel refactored into EntityExport and EntityExport.Settings.
+- EntityExport.Settings.Attributes added.
 ### is.codion.swing.framework.model
 - SwingEntityConditionModelFactory renamed SwingEntityConditions.
 ### is.codion.swing.framework.ui
@@ -59,6 +60,7 @@ Codion Change Log
 - EntityTableExportPanel bug fixed, unselected foreign key attributes with selected children can now be moved.
 - EntityTableExportPanel, intermittent index out of bounds during multiple item move bug fixed, indexes now sorted, order not guaranteed by getSelectionPaths().
 - EntityTableExportModel foreign key nodes now indicate the number of selected child attributes.
+- EntityTableExportModel.sortAttributes() ClassCastException bug fixed.
 
 ## 0.18.60
 ### is.codion.common.reactive